Merge from tencent/bk-ci

+++ src/backend/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/service/NodeService.kt
@@ -28,9 +28,11 @@ package com.tencent.devops.environment.service
 
 import com.tencent.devops.common.api.exception.OperationException
 import com.tencent.devops.common.api.util.HashUtil
+import com.tencent.devops.common.auth.api.BkAuthPermission
 import com.tencent.devops.environment.dao.EnvNodeDao
 import com.tencent.devops.environment.dao.NodeDao
 import com.tencent.devops.environment.dao.thirdPartyAgent.ThirdPartyAgentDao
+import com.tencent.devops.environment.permission.EnvironmentPermissionService
 import com.tencent.devops.environment.pojo.NodeBaseInfo
 import com.tencent.devops.environment.pojo.NodeWithPermission
 import com.tencent.devops.environment.pojo.enums.NodeStatus
@@ -51,27 +53,60 @@ class NodeService @Autowired constructor(
     private val nodeDao: NodeDao,
     private val envNodeDao: EnvNodeDao,
     private val thirdPartyAgentDao: ThirdPartyAgentDao,
-    private val slaveGatewayService: SlaveGatewayService
+    private val slaveGatewayService: SlaveGatewayService,
+    private val environmentPermissionService: EnvironmentPermissionService
 ) {
     fun deleteNodes(userId: String, projectId: String, nodeHashIds: List<String>) {
         val nodeLongIds = nodeHashIds.map { HashUtil.decodeIdToLong(it) }
+        val canDeleteNodeIds =
+            environmentPermissionService.listNodeByPermission(userId, projectId, BkAuthPermission.DELETE)
         val existNodeList = nodeDao.listByIds(dslContext, projectId, nodeLongIds)
         if (existNodeList.isEmpty()) {
             return
         }
         val existNodeIdList = existNodeList.map { it.nodeId }
 
+        val unauthorizedNodeIds = existNodeIdList.filterNot { canDeleteNodeIds.contains(it) }
+        if (unauthorizedNodeIds.isNotEmpty()) {
+            throw OperationException("没有删除节点的权限，节点ID：[${unauthorizedNodeIds.map { HashUtil.encodeLongId(it) }.joinToString(",")}]")
+        }
+
         dslContext.transaction { configuration ->
             val context = DSL.using(configuration)
             nodeDao.batchDeleteNode(context, projectId, existNodeIdList)
             envNodeDao.deleteByNodeIds(context, existNodeIdList)
+            existNodeIdList.forEach {
+                environmentPermissionService.deleteNode(projectId, it)
+            }
         }
+    }
+
+    fun hasCreatePermission(userId: String, projectId: String): Boolean {
+        return environmentPermissionService.checkNodePermission(userId, projectId, BkAuthPermission.CREATE)
     }
 
     fun list(userId: String, projectId: String): List<NodeWithPermission> {
         val nodeRecordList = nodeDao.listNodes(dslContext, projectId)
         if (nodeRecordList.isEmpty()) {
             return emptyList()
+        }
+
+        val permissionMap = environmentPermissionService.listNodeByPermissions(userId, projectId,
+            setOf(BkAuthPermission.USE, BkAuthPermission.EDIT, BkAuthPermission.DELETE))
+        val canUseNodeIds = if (permissionMap.containsKey(BkAuthPermission.USE)) {
+            permissionMap[BkAuthPermission.USE]!!.map { HashUtil.decodeIdToLong(it) }
+        } else {
+            emptyList()
+        }
+        val canEditNodeIds = if (permissionMap.containsKey(BkAuthPermission.EDIT)) {
+            permissionMap[BkAuthPermission.EDIT]!!.map { HashUtil.decodeIdToLong(it) }
+        } else {
+            emptyList()
+        }
+        val canDeleteNodeIds = if (permissionMap.containsKey(BkAuthPermission.DELETE)) {
+            permissionMap[BkAuthPermission.DELETE]!!.map { HashUtil.decodeIdToLong(it) }
+        } else {
+            emptyList()
         }
 
         val thirdPartyAgentNodeIds = nodeRecordList.filter { it.nodeType == NodeType.THIRDPARTY.name }
@@ -89,27 +124,27 @@ class NodeService @Autowired constructor(
             val nodeStringId = NodeStringIdUtils.getNodeStringId(it)
             NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName)
             NodeWithPermission(
-                HashUtil.encodeLongId(it.nodeId),
-                nodeStringId,
-                it.nodeName,
-                it.nodeIp,
-                it.nodeStatus,
-                getAgentStatus(it),
-                it.nodeType,
-                it.osName,
-                it.createdUser,
-                it.operator,
-                it.bakOperator,
-                true,
-                true,
-                true,
-                gatewayShowName,
-                NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName),
-                if (null == it.createdTime) "" else
+                nodeHashId = HashUtil.encodeLongId(it.nodeId),
+                nodeId = nodeStringId,
+                name = it.nodeName,
+                ip = it.nodeIp,
+                nodeStatus = it.nodeStatus,
+                agentStatus = getAgentStatus(it),
+                nodeType = it.nodeType,
+                osName = it.osName,
+                createdUser = it.createdUser,
+                operator = it.operator,
+                bakOperator = it.bakOperator,
+                canUse = canUseNodeIds.contains(it.nodeId),
+                canEdit = canEditNodeIds.contains(it.nodeId),
+                canDelete = canDeleteNodeIds.contains(it.nodeId),
+                gateway = gatewayShowName,
+                displayName = NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName),
+                createTime = if (null == it.createdTime) "" else
                     DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").format(it.createdTime),
-                if (null == it.lastModifyTime) "" else
+                lastModifyTime = if (null == it.lastModifyTime) "" else
                     DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").format(it.lastModifyTime),
-                it.lastModifyUser ?: ""
+                lastModifyUser = it.lastModifyUser ?: ""
             )
         }
     }
@@ -119,6 +154,25 @@ class NodeService @Autowired constructor(
         val nodeRecordList = nodeDao.listAllByIds(dslContext, projectId, nodeIds)
         if (nodeRecordList.isEmpty()) {
             return emptyList()
+        }
+
+        val permissionMap = environmentPermissionService.listNodeByPermissions(userId, projectId,
+            setOf(BkAuthPermission.USE, BkAuthPermission.EDIT, BkAuthPermission.DELETE))
+
+        val canUseNodeIds = if (permissionMap.containsKey(BkAuthPermission.USE)) {
+            permissionMap[BkAuthPermission.USE]!!.map { HashUtil.decodeIdToLong(it) }
+        } else {
+            emptyList()
+        }
+        val canEditNodeIds = if (permissionMap.containsKey(BkAuthPermission.EDIT)) {
+            permissionMap[BkAuthPermission.EDIT]!!.map { HashUtil.decodeIdToLong(it) }
+        } else {
+            emptyList()
+        }
+        val canDeleteNodeIds = if (permissionMap.containsKey(BkAuthPermission.DELETE)) {
+            permissionMap[BkAuthPermission.DELETE]!!.map { HashUtil.decodeIdToLong(it) }
+        } else {
+            emptyList()
         }
 
         val thirdPartyAgentNodeIds = nodeRecordList.filter { it.nodeType == NodeType.THIRDPARTY.name }
@@ -134,27 +188,27 @@ class NodeService @Autowired constructor(
             }
             val nodeStringId = NodeStringIdUtils.getNodeStringId(it)
             NodeWithPermission(
-                HashUtil.encodeLongId(it.nodeId),
-                nodeStringId,
-                it.nodeName,
-                it.nodeIp,
-                NodeStatus.getStatusName(it.nodeStatus),
-                getAgentStatus(it),
-                NodeType.getTypeName(it.nodeType),
-                it.osName,
-                it.createdUser,
-                it.operator,
-                it.bakOperator,
-                true,
-                true,
-                true,
-                gatewayShowName,
-                NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName),
-                if (null == it.createdTime) "" else
+                nodeHashId = HashUtil.encodeLongId(it.nodeId),
+                nodeId = nodeStringId,
+                name = it.nodeName,
+                ip = it.nodeIp,
+                nodeStatus = NodeStatus.getStatusName(it.nodeStatus),
+                agentStatus = getAgentStatus(it),
+                nodeType = NodeType.getTypeName(it.nodeType),
+                osName = it.osName,
+                createdUser = it.createdUser,
+                operator = it.operator,
+                bakOperator = it.bakOperator,
+                canUse = canUseNodeIds.contains(it.nodeId),
+                canEdit = canEditNodeIds.contains(it.nodeId),
+                canDelete = canDeleteNodeIds.contains(it.nodeId),
+                gateway = gatewayShowName,
+                displayName = NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName),
+                createTime = if (null == it.createdTime) "" else
                     DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").format(it.createdTime),
-                if (null == it.lastModifyTime) "" else
+                lastModifyTime = if (null == it.lastModifyTime) "" else
                     DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").format(it.lastModifyTime),
-                it.lastModifyUser ?: ""
+                lastModifyUser = it.lastModifyUser ?: ""
             )
         }
     }
@@ -164,19 +218,19 @@ class NodeService @Autowired constructor(
         return nodeRecords.map {
             val nodeStringId = NodeStringIdUtils.getNodeStringId(it)
             NodeBaseInfo(
-                HashUtil.encodeLongId(it.nodeId),
-                nodeStringId,
-                it.nodeName,
-                it.nodeIp,
-                it.nodeStatus,
-                getAgentStatus(it),
-                it.nodeType,
-                it.osName,
-                it.createdUser,
-                it.operator,
-                it.bakOperator,
-                "",
-                NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName)
+                nodeHashId = HashUtil.encodeLongId(it.nodeId),
+                nodeId = nodeStringId,
+                name = it.nodeName,
+                ip = it.nodeIp,
+                nodeStatus = it.nodeStatus,
+                agentStatus = getAgentStatus(it),
+                nodeType = it.nodeType,
+                osName = it.osName,
+                createdUser = it.createdUser,
+                operator = it.operator,
+                bakOperator = it.bakOperator,
+                gateway = "",
+                displayName = NodeStringIdUtils.getRefineDisplayName(nodeStringId, it.displayName)
             )
         }
     }
@@ -190,11 +244,13 @@ class NodeService @Autowired constructor(
     fun updateDisplayName(userId: String, projectId: String, nodeHashId: String, displayName: String) {
         val nodeId = HashUtil.decodeIdToLong(nodeHashId)
         nodeDao.get(dslContext, projectId, nodeId) ?: throw NotFoundException("node not found")
+        if (!environmentPermissionService.checkNodePermission(userId, projectId, BkAuthPermission.EDIT)) {
+            throw OperationException("No Permission")
+        }
         checkDisplayName(projectId, nodeId, displayName)
-
         dslContext.transaction { configuration ->
             val context = DSL.using(configuration)
-            nodeDao.updateDisplayName(context, nodeId, displayName)
+            nodeDao.updateDisplayName(dslContext = context, nodeId = nodeId, nodeName = displayName, userId = userId)
         }
     }
 }
